% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{month_diff}
\alias{month_diff}
\title{month_diff}
\usage{
month_diff(month2, month1, numeric = F)
}
\arguments{
\item{month2}{character}

\item{month1}{character}

\item{numeric}{logical}
}
\description{
month_diff
}
