% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{time_diff}
\alias{time_diff}
\title{time_diff}
\usage{
time_diff(time2, time1, unit = "mins", round_n = 2)
}
\arguments{
\item{time2}{Date}

\item{time1}{Date}

\item{unit}{character}
}
\description{
time_diff
}
