% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{week_diff}
\alias{week_diff}
\title{week_diff}
\usage{
week_diff(week2, week1, numeric = F)
}
\arguments{
\item{week2}{character}

\item{week1}{character}

\item{numeric}{logical}
}
\description{
week_diff
}
